+++ linghe/utils/mul.py
@@ -93,6 +93,7 @@ def triton_inplace_scale(x, scale):
 
 @triton.jit
 def batch_scale_kernel(input_ptrs, size_ptr, scale, 
+                       DT: tl.constexpr,
                        B: tl.constexpr,
                        ZERO: tl.constexpr,):
     tid = tl.program_id(axis=0)
@@ -100,7 +101,10 @@ def batch_scale_kernel(input_ptrs, size_ptr, scale,
     T = tl.num_programs(axis=1)
 
     size = tl.load(size_ptr + tid)
-    input_ptr = tl.load(input_ptrs + tid).to(tl.pointer_type(tl.float32))
+    if DT == 0:
+        input_ptr = tl.load(input_ptrs + tid).to(tl.pointer_type(tl.float32))
+    else:
+        input_ptr = tl.load(input_ptrs + tid).to(tl.pointer_type(tl.bfloat16))
     t = tl.cdiv(size, B * T)
     offs = bid.to(tl.int64) * t * B + tl.arange(0, B)
     for i in range(t):
@@ -123,7 +127,12 @@ def triton_batch_scale(xs, scale):
     Returns:
         xs
     """
-    assert all([x.is_contiguous() and x.dtype == torch.float32 for x in xs])
+    if len(xs) == 0:
+        return
+    dtype = xs[0].dtype
+    assert dtype in (torch.float32, torch.bfloat16)
+    assert all([x.is_contiguous() and x.dtype == dtype for x in xs])
+
 
     device = xs[0].device
     sizes = torch.tensor([x.numel() for x in xs], 
@@ -131,6 +140,7 @@ def triton_batch_scale(xs, scale):
     ptrs = torch.tensor([x.data_ptr() for x in xs], 
                         dtype=torch.int64).cuda(device, non_blocking=True)
 
+    DT = 0 if dtype == torch.float32 else 1
     T = 256
     tensor_count = len(xs)
     B = 512
@@ -140,6 +150,7 @@ def triton_batch_scale(xs, scale):
         ptrs,
         sizes,
         scale,
+        DT,
         B,
         ZERO,
         num_stages=2,